+++ tests/testsuite/main.rs
@@ -1,4 +1,6 @@
 extern crate elba;
+#[macro_use]
+extern crate indexmap;
 extern crate itertools;
 #[macro_use]
 extern crate lazy_static;